implementing log tracing between services on product-api

+++ product-api/src/main/java/br/com/trilhabackend/productapi/modules/product/service/ProductService.java
@@ -8,9 +8,12 @@ import br.com.trilhabackend.productapi.modules.product.model.Product;
 import br.com.trilhabackend.productapi.modules.product.repository.ProductRepository;
 import br.com.trilhabackend.productapi.modules.sales.client.SalesClient;
 import br.com.trilhabackend.productapi.modules.sales.dto.SalesConfirmationDTO;
+import br.com.trilhabackend.productapi.modules.sales.dto.SalesProductResponse;
 import br.com.trilhabackend.productapi.modules.sales.enums.SalesStatus;
 import br.com.trilhabackend.productapi.modules.sales.rabbitmq.SalesConfirmationSender;
 import br.com.trilhabackend.productapi.modules.supplier.service.SupplierService;
+import com.fasterxml.jackson.databind.ObjectMapper;
+
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,11 +23,14 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static br.com.trilhabackend.productapi.config.RequestUtil.getCurrentRequest;
 import static org.springframework.util.ObjectUtils.isEmpty;
 @Slf4j
 @Service
 public class ProductService {
     private static final Integer ZERO = 0;
+    private static final String TRANSACTION_ID = "transactionid";
+    private static final String SERVICE_ID = "serviceid";
 
     @Autowired
     private ProductRepository productRepository;
@@ -128,8 +134,15 @@ public class ProductService {
     public Boolean existsBySupplierId(Integer supplierId) {
         return productRepository.existsByCategoryId(supplierId);
     }
-    public SuccessResponse delete (Integer id) {
+    public SuccessResponse delete(Integer id) {
         validateInformedId(id);
+        if (!productRepository.existsById(id)) {
+            throw new ValidationException("The product does not exists.");
+        }
+        var sales = getSalesByProductId(id);
+        if (!isEmpty(sales.getSalesIds())) {
+            throw new ValidationException("The product cannot be deleted. There are sales for it.");
+        }
         productRepository.deleteById(id);
         return SuccessResponse.create("The product was deleted.");
     }
@@ -145,7 +158,7 @@ public class ProductService {
             updateStock(product);
         } catch (Exception ex) {
             log.error("Error while trying to update stock for message with error: {}", ex.getMessage(), ex);
-            var rejectedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatus.REJECTED);
+            var rejectedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatus.REJECTED, product.getTransactionid());
             salesConfirmationSender.sendSalesConfirmationMessage(rejectedMessage);
         }
     }
@@ -181,7 +194,7 @@ public class ProductService {
                 });
         if(!isEmpty(productsForUpdate)) {
             productRepository.saveAll(productsForUpdate);
-            var approvedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatus.APPROVED);
+            var approvedMessage = new SalesConfirmationDTO(product.getSalesId(), SalesStatus.APPROVED, product.getTransactionid());
             salesConfirmationSender.sendSalesConfirmationMessage(approvedMessage);
         }
     }
@@ -196,26 +209,48 @@ public class ProductService {
 
     public ProductSalesResponse findProductSales(Integer id) {
         var product = findById(id);
+        var sales = getSalesByProductId(product.getId());
+        return ProductSalesResponse.of(product, sales.getSalesIds());
+    }
+
+    private SalesProductResponse getSalesByProductId (Integer productId) {
         try {
-            var sales = salesClient
-                    .findSalesByProductId(product.getId())
+            var currentRequest = getCurrentRequest();
+            var transactionid = currentRequest.getHeader(TRANSACTION_ID);
+            var serviceid = currentRequest.getAttribute(SERVICE_ID);
+            log.info("Sending GET request to orders by productId with data {} | [transactionID: {} | serviceID: {}]",
+                    productId, transactionid, serviceid);
+            var response = salesClient
+                    .findSalesByProductId(productId)
                     .orElseThrow(() -> new ValidationException("The sales was not found by this product."));
-            return ProductSalesResponse.of(product, sales.getSalesIds());
+            log.info("Receiving response from orders by productId with data {} | [transactionID: {} | serviceID: {}]",
+                    new ObjectMapper().writeValueAsString(response), transactionid, serviceid);
+            return response;
         } catch (Exception ex) {
-            throw new ValidationException("There was an error trying to get the product's sales.");
+            throw new ValidationException("The sales could not be found.");
         }
     }
 
     public SuccessResponse checkProductsStock(ProductCheckStockRequest request) {
-        if (isEmpty(request) || isEmpty(request.getProducts())) {
-            throw new ValidationException("The requested data and products must be informed.");
+        try {
+            var currentRequest = getCurrentRequest();
+            var transactionid = currentRequest.getHeader(TRANSACTION_ID);
+            var serviceid = currentRequest.getAttribute(SERVICE_ID);
+            log.info("Request to POST product stock with data {} | [transactionID: {} | serviceID: {}]",
+                    new ObjectMapper().writeValueAsString(request), transactionid, serviceid);
+            if (isEmpty(request) || isEmpty(request.getProducts())) {
+                throw new ValidationException("The requested data and products must be informed.");
+            }
+            System.out.println(request.getProducts());
+            request
+                    .getProducts()
+                    .forEach(this::validateStock);
+            var response = SuccessResponse.create("The stock is ok!");
+            log.info("Response to POST product stock with data {} | [transactionID: {} | serviceID: {}]", new ObjectMapper().writeValueAsString(response), transactionid, serviceid);
+            return response;
+        } catch (Exception ex) {
+            throw new ValidationException(ex.getMessage());
         }
-        System.out.println(request.getProducts());
-        request
-                .getProducts()
-                .forEach(this::validateStock);
-        return SuccessResponse.create("The stock is ok!");
-
     }
     private void validateStock(ProductQuantityDTO productQuantity) {
         if (isEmpty(productQuantity.getProductId()) || isEmpty(productQuantity.getQuantity())) {
